chore(vc): update utils
